Change player object activation
The player object itself is no longer disabled when it is no its turn but an inner controller object, if it exists, is disabled instead.
This is to allow all players to continue interacting with the game even when it is not their turn, as well as handling several local players.