Remove the half tile increase to ability range
The increase was used to make more natural looking circles but it breaks the logic that the range should be the distance between two tile centers and not between the tile center and the other tile border.
An alternative would be to have range as float and add the half tile directly in the ability definition in the content files.