Reset the main game scene to include only the minimum required setup
The game scene contained an already initialized game even when it would be recreated on start.
This meant the scene often changed due to changes on other objects, keeping the scene to a minimum setup should reduce the need to commit changes to it.
The scene was also used to test changes, this should be done with a local copy of the scene.