Update the command input handling
Refactors updates to have a single flow for the command execution.
Uses constants for left and right click values for Unity input.
Resets ability target area color when the ability changes.
Notifies about ability action points and range when failing to execute a command.
Checks for null ability in AbilityCastView.